Add classic "Hello, world!" C++ program
This commit adds the quintessential "Hello, world!" C++ program to the repository. With just a few lines of code, this simple yet powerful program demonstrates the essence of C++ programming. Through the magic of the iostream library, the program prints the timeless greeting "Hello, world!" to the console, marking the beginning of countless journeys into the realm of software development. As developers, we often start our coding adventures with this humble program, laying the foundation for understanding basic syntax and the flow of control in C++. So, let this commit serve as a tribute to the tradition of learning and exploration in the vast landscape of programming.