Add C++ program to check a number is prime or not
This commit introduces a C++ program named "Prime Number Checker" that determines whether a given integer is a prime number or not. The program utilizes a class called "number" with methods for setting data and checking whether the number is prime. It takes user input, sets the data, and then checks whether the number is prime or not using a simple algorithm. This commit provides a basic implementation of the prime number checking functionality.