Add C++ program to calculate power of a number
This commit adds a C++ program that calculates the power of a number using the pow function from the <cmath> library. The program prompts the user to enter two integers, m and n, representing the base and exponent respectively. It then computes and displays m raised to the power of n. This code can be useful for various mathematical applications where exponentiation is required.