Add C++ program to calculate Fibonacci series
This commit implements a C++ program to calculate the Fibonacci series. The Fibonacci function recursively calculates the Fibonacci value for a given index. In the main function, it prompts the user to enter the number of Fibonacci numbers to be generated and then iterates through the series, printing each Fibonacci number. The program is designed to handle non-negative integer inputs.