Added C++ code for copy constructor
This commit introduces a C++ program that defines a Point class representing coordinates in a two-dimensional space. It includes a constructor to initialize the Point object with given coordinates, as well as a copy constructor for creating a new Point object from an existing one. The main function demonstrates the usage of the Point class by creating two Point objects, p1 and p2, and copying the values of p1 into p2. Finally, it prints out the coordinates of both points.