Add Tower of Hanoi implementation in C++
The program takes input for the number of plates to be exchanged and recursively solves the Tower of Hanoi problem, printing the sequence of moves required to move all the plates from one rod to another following the rules of the puzzle. The main function prompts the user to input the number of plates and then calls the tower_of_Hanoi function to perform the necessary operations.