Added C++ program to show object pointer
This commit introduces a C++ program that demonstrates the implementation of a student class. The student class has private data members for storing a float result and a public member ID. The constructor initializes these members using parameters passed during object creation. Additionally, the class includes a method show() to display the ID and result of the student. In the main() function, an instance of the student class is created with specific values for result and ID. A pointer to this instance is then used to invoke the show() method, demonstrating basic usage of classes and pointers in C++. This commit aims to provide a clear example of object-oriented principles and pointer usage in C++.